move nonce setter base to a common place
